Moving outside code to terraform root module and removing symbolic links
