--- requirements.txt
@@ -1,42 +1,0 @@
-attrs
-brotlipy
-certifi
-cffi
-chardet
-click
-cryptography
-dnspython
-Flask
-gunicorn
-idna
-importlib-metadata
-iniconfig
-itsdangerous
-Jinja2
-joblib
-loguru
-MarkupSafe
-more-itertools
-packaging
-pip
-pluggy
-py
-pycparser
-pymongo
-pyOpenSSL
-pyparsing
-PySocks
-pytest
-python-dateutil
-python-dotenv
-pytz
-requests
-setuptools
-six
-threadpoolctl
-toml
-tomli
-urllib3
-Werkzeug
-wheel
-zipp
