--- terraform/project_dependencies/common/python/dao
@@ -1,1 +1,0 @@
-../../../../dao
