--- terraform/project_dependencies/common/python/exceptions
@@ -1,1 +1,0 @@
-../../../../exceptions
