+++ terraform/project_dependencies/common/python/exceptions/schema_validation_exception.py
