--- terraform/project_dependencies/common/python/model
@@ -1,1 +1,0 @@
-../../../../model
