+++ terraform/project_dependencies/common/python/model/base_model.py
