--- terraform/project_dependencies/common/python/repository
@@ -1,1 +1,0 @@
-../../../../repository
