+++ terraform/project_dependencies/common/python/repository/cache_repository.py
@@ -1,0 +1,23 @@
+from datetime import datetime
+
+from terraform.project_dependencies.common.python.dao.cache_dao import CacheDao
+from terraform.project_dependencies.common.python.model import Cache
+
+
+class CacheRepository:
+    def __init__(self):
+        self._cache_dao = CacheDao()
+
+    def get_valid_cache(self) -> Cache:
+        last_available_cache = self._cache_dao.select_one({'is_valid': True})
+        if last_available_cache is not None:
+            if datetime.fromtimestamp(
+                    datetime.timestamp(datetime.now()) - last_available_cache.timestamp).time().hour > 1:
+                last_available_cache.is_valid = False
+                self._cache_dao.update(last_available_cache)
+            return last_available_cache
+        else:
+            return Cache()
+
+    def update_cache(self):
+        self._cache_dao.insert(Cache())
