+++ terraform/project_dependencies/common/python/repository/user_repository.py
@@ -1,0 +1,10 @@
+from terraform.project_dependencies.common.python.dao.user_dao import UserDao
+from terraform.project_dependencies.common.python.model.user import User
+
+
+class UserRepository:
+    def __init__(self):
+        self.user_dao = UserDao()
+
+    def check_user_login(self, username: str, password: str) -> [User, None]:
+        return self.user_dao.select_one({'username': username, 'password': password})
