--- terraform/project_dependencies/common/python/util
@@ -1,1 +1,0 @@
-../../../../util
