+++ terraform/project_dependencies/common/python/util/singleton.py
