--- terraform/project_dependencies/currency
@@ -1,1 +1,0 @@
-../../currency/
