+++ terraform/project_dependencies/currency/currency.py
@@ -1,0 +1,26 @@
+from typing import Any, Optional
+
+import requests
+
+from terraform.project_dependencies.common.python import constants
+
+from terraform.project_dependencies.common.python.logger import Logger
+from terraform.project_dependencies.common.python.util.response import Response, CommonHeaders
+
+
+def get_currencies_lambda_handler(event: dict, context: Optional[dict]) -> dict[str, Any]:
+    return get_currencies(event, context)
+
+
+def get_currencies(event: dict, context: Optional[dict]) -> dict[str, Any]:
+    Logger.d("Passed event: %s", event)
+    if context:
+        Logger.i("Passed context: %s", context)
+    currency_list_res = requests.get(url=constants.currencies_url)
+    headers = {
+        CommonHeaders.ContentType.value: "application/json"
+    }
+    body = {
+        "currencies ": str(currency_list_res.content, "utf-8")
+    }
+    return Response(status_code=200, headers=headers, body=body).to_dict()
