+++ terraform/project_dependencies/currency/requirements.txt
