--- terraform/project_dependencies/history
@@ -1,1 +1,0 @@
-../../history
