--- util/providers/base_database_provider.py
@@ -1,15 +1,0 @@
-import abc
-
-from pymongo import MongoClient
-
-from config import Config
-
-
-class BaseDatabaseProvider(abc.ABC):
-
-    @abc.abstractmethod
-    def get_connection(self) -> MongoClient:
-        pass
-
-    def get_database(self):
-        return self.get_connection().get_database(Config.DB_NAME.value)
